+++ tests/Integration/Execution/MutationExecutor/BelongsToManyTest.php
@@ -160,10 +160,10 @@ class BelongsToManyTest extends DBTestCase
                     'id' => '1',
                     'users' => [
                         [
-                            'id' => '1',
+                            'id' => '2',
                         ],
                         [
-                            'id' => '2',
+                            'id' => '1',
                         ],
                     ],
                 ],
@@ -178,12 +178,14 @@ class BelongsToManyTest extends DBTestCase
             createRole(input: {
                 name: "foobar"
                 users: {
-                    create: [{
-                        name: "bar"
-                    },
-                    {
-                        name: "foo"
-                    }]
+                    create: [
+                        {
+                            name: "bar"
+                        },
+                        {
+                            name: "foo"
+                        }
+                    ]
                 }
             }) {
                 id
@@ -218,14 +220,16 @@ class BelongsToManyTest extends DBTestCase
                 id: 1
                 name: "is_user"
                 users: {
-                    upsert: [{
-                        id: 10
-                        name: "user1"
-                    },
-                    {
-                        id: 20
-                        name: "user2"
-                    }]
+                    upsert: [
+                        {
+                            id: 10
+                            name: "user1"
+                        },
+                        {
+                            id: 20
+                            name: "user2"
+                        }
+                    ]
                 }
             }) {
                 id
@@ -268,12 +272,14 @@ class BelongsToManyTest extends DBTestCase
             upsertRole(input: {
                 name: "is_user"
                 users: {
-                    upsert: [{
-                        name: "user1"
-                    },
-                    {
-                        name: "user2"
-                    }]
+                    upsert: [
+                        {
+                            name: "user1"
+                        },
+                        {
+                            name: "user2"
+                        }
+                    ]
                 }
             }) {
                 id
@@ -320,9 +326,7 @@ GRAPHQL
             createRole(input: {
                 name: "foobar"
                 users: {
-                    connect: [
-                        1,2
-                    ]
+                    connect: [1, 2]
                 }
             }) {
                 id
@@ -364,9 +368,7 @@ GRAPHQL
                 id: 1
                 name: "foobar"
                 users: {
-                    connect: [
-                        1,2
-                    ]
+                    connect: [1, 2]
                 }
             }) {
                 id
@@ -409,12 +411,14 @@ GRAPHQL
                 id: 1
                 name: "is_user"
                 users: {
-                    create: [{
-                        name: "user1"
-                    },
-                    {
-                        name: "user2"
-                    }]
+                    create: [
+                        {
+                            name: "user1"
+                        },
+                        {
+                            name: "user2"
+                        }
+                    ]
                 }
             }) {
                 id
@@ -501,14 +505,16 @@ GRAPHQL
                 id: 1
                 name: "is_user"
                 users: {
-                    upsert: [{
-                        id: 1
-                        name: "user1"
-                    },
-                    {
-                        id: 2
-                        name: "user2"
-                    }]
+                    upsert: [
+                        {
+                            id: 1
+                            name: "user1"
+                        },
+                        {
+                            id: 2
+                            name: "user2"
+                        }
+                    ]
                 }
             }) {
                 id
@@ -566,19 +572,19 @@ GRAPHQL
                 factory(User::class, 2)->create()
             );
 
-        $this->graphQL(/** @lang GraphQL */ "
+        $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
             ${action}Role(input: {
                 id: 1
-                name: \"is_user\"
+                name: "is_user"
                 users: {
                     ${action}: [{
                         id: 1
-                        name: \"user1\"
+                        name: "user1"
                     },
                     {
                         id: 2
-                        name: \"user2\"
+                        name: "user2"
                     }]
                 }
             }) {
@@ -590,7 +596,8 @@ GRAPHQL
                 }
             }
         }
-        ")->assertJson([
+GRAPHQL
+        )->assertJson([
             'data' => [
                 "${action}Role" => [
                     'id' => '1',
@@ -629,11 +636,11 @@ GRAPHQL
                 factory(User::class, 2)->create()
             );
 
-        $this->graphQL(/** @lang GraphQL */ "
+        $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
             ${action}Role(input: {
                 id: 1
-                name: \"is_user\"
+                name: "is_user"
                 users: {
                     delete: [1]
                 }
@@ -645,7 +652,8 @@ GRAPHQL
                 }
             }
         }
-        ")->assertJson([
+GRAPHQL
+        )->assertJson([
             'data' => [
                 "${action}Role" => [
                     'id' => '1',
@@ -681,7 +689,7 @@ GRAPHQL
                 factory(User::class)->create()
             );
 
-        $this->graphQL(/** @lang GraphQL */ "
+        $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
             ${action}Role(input: {
                 id: 1
@@ -696,16 +704,17 @@ GRAPHQL
                 }
             }
         }
-        ")->assertJson([
+GRAPHQL
+        )->assertJson([
             'data' => [
                 "${action}Role" => [
                     'id' => '1',
                     'users' => [
                         [
-                            'id' => '1',
+                            'id' => '2',
                         ],
                         [
-                            'id' => '2',
+                            'id' => '1',
                         ],
                     ],
                 ],
@@ -730,12 +739,12 @@ GRAPHQL
                 factory(User::class)->create()
             );
 
-        $this->graphQL(/** @lang GraphQL */ "
+        $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
             ${action}Role(input: {
                 id: 1
                 users: {
-                    sync: [1,2]
+                    sync: [1, 2]
                 }
             }) {
                 id
@@ -745,16 +754,17 @@ GRAPHQL
                 }
             }
         }
-        ")->assertJson([
+GRAPHQL
+        )->assertJson([
             'data' => [
                 "${action}Role" => [
                     'id' => '1',
                     'users' => [
                         [
-                            'id' => '1',
+                            'id' => '2',
                         ],
                         [
-                            'id' => '2',
+                            'id' => '1',
                         ],
                     ],
                 ],
@@ -778,7 +788,7 @@ GRAPHQL
                 factory(User::class, 2)->create()
             );
 
-        $this->graphQL(/** @lang GraphQL */ "
+        $this->graphQL(/** @lang GraphQL */ <<<GRAPHQL
         mutation {
             ${action}Role(input: {
                 id: 1
@@ -792,7 +802,8 @@ GRAPHQL
                 }
             }
         }
-        ")->assertJson([
+GRAPHQL
+        )->assertJson([
             'data' => [
                 "${action}Role" => [
                     'id' => '1',
@@ -822,9 +833,7 @@ GRAPHQL
             createRole(input: {
                 name: "foobar"
                 users: {
-                    sync: [
-                        1,2
-                    ]
+                    sync: [1, 2]
                 }
             }) {
                 id
@@ -862,9 +871,7 @@ GRAPHQL
                 id: 1
                 name: "foobar"
                 users: {
-                    sync: [
-                        1,2
-                    ]
+                    sync: [1, 2]
                 }
             }) {
                 id
@@ -942,17 +949,17 @@ GRAPHQL
         $role2 = factory(Role::class)->create();
         $user->roles()->attach($role2);
 
-        $metaText = Lorem::sentence();
+        $meta = Lorem::sentence();
 
         $this->graphQL(/** @lang GraphQL */ '
-        mutation {
+        mutation ($meta: String) {
             pivotsUpdateUser(input: {
                 id: 1,
                 roles: {
                     sync: [
                         {
                             id: 1,
-                            meta: "'.$metaText.'"
+                            meta: $meta
                         },
                         {
                             id: 2
@@ -968,20 +975,22 @@ GRAPHQL
                 }
             }
         }
-        ')->assertJson([
+        ', [
+            'meta' => $meta,
+        ])->assertJson([
             'data' => [
                 'pivotsUpdateUser' => [
                     'roles' => [
                         [
-                            'id' => 1,
-                            'pivot' => [
-                                'meta' => $metaText,
-                            ],
-                        ],
-                        [
                             'id' => 2,
                             'pivot' => [
                                 'meta' => null,
+                            ],
+                        ],
+                        [
+                            'id' => 1,
+                            'pivot' => [
+                                'meta' => $meta,
                             ],
                         ],
                     ],
@@ -997,17 +1006,17 @@ GRAPHQL
         $role2 = factory(Role::class)->create();
         $user->roles()->attach($role2);
 
-        $metaText = Lorem::sentence();
+        $meta = Lorem::sentence();
 
         $this->graphQL(/** @lang GraphQL */ '
-        mutation {
+        mutation ($meta: String) {
             pivotsUpdateUser(input: {
                 id: 1,
                 roles: {
                     syncWithoutDetaching: [
                         {
                             id: 1,
-                            meta: "'.$metaText.'"
+                            meta: $meta
                         }
                     ]
                 },
@@ -1019,18 +1028,20 @@ GRAPHQL
                 }
             }
         }
-        ')->assertJson([
+        ', [
+            'meta' => $meta,
+        ])->assertJson([
             'data' => [
                 'pivotsUpdateUser' => [
                     'roles' => [
                         [
                             'pivot' => [
-                                'meta' => $metaText,
+                                'meta' => null,
                             ],
                         ],
                         [
                             'pivot' => [
-                                'meta' => null,
+                                'meta' => $meta,
                             ],
                         ],
                     ],
@@ -1041,20 +1052,20 @@ GRAPHQL
 
     public function testCanConnectUserWithRoleAndPivotMetaByUsingConnect(): void
     {
-        $user = factory(User::class)->create();
-        $role = factory(Role::class)->create();
+        factory(User::class)->create();
+        factory(Role::class)->create();
 
-        $metaText = Lorem::sentence();
+        $meta = Lorem::sentence();
 
         $this->graphQL(/** @lang GraphQL */ '
-        mutation {
+        mutation ($meta: String) {
             pivotsUpdateUser(input: {
                 id: 1,
                 roles: {
                     connect: [
                         {
                             id: 1,
-                            meta: "'.$metaText.'"
+                            meta: $meta
                         }
                     ]
                 },
@@ -1066,13 +1077,15 @@ GRAPHQL
                 }
             }
         }
-        ')->assertJson([
+        ', [
+            'meta' => $meta,
+        ])->assertJson([
             'data' => [
                 'pivotsUpdateUser' => [
                     'roles' => [
                         [
                             'pivot' => [
-                                'meta' => $metaText,
+                                'meta' => $meta,
                             ],
                         ],
                     ],